refactor: move store to its own folder
