+++ jsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "module": "es6",
+    "module": "commonJS",
     "allowSyntheticDefaultImports": true,
     "jsx": "react"
   }
