+++ src/modules/data/store/types.js
@@ -1,0 +1,6 @@
+export class Group {
+  name = '';
+  describe = '';
+  countUnit = '';
+  samples = [];
+}
